pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "torchdtcc"
version = "0.0.1"
description = "Deep Temporal Contrastive Clustering for time series"
authors = [
  { name="Zsombor Kalotay", email="zsombor.kalotay@gmail.ch" }
]
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
dependencies = [
    "torch>=1.8",
    "numpy>=1.19",
    "tslearn>=0.5.2",
    "scikit-learn>=0.22",
    "tqdm>=4.45",
    "pyyaml>=5.1",
    "scipy>=1.5"
]

[project.optional-dependencies]
dev = [
    "jupyter",
    "pytest",
    "black",
    "flake8"
]

[tool.setuptools]
package-dir = {"" = "torchdtcc"}
packages = ["torchdtcc"]
include-package-data = true

[tool.setuptools.package-data]
torchdtcc = ["*.py"]